requirements.txt
# JARVIS v2.0 - Dependências do Projeto
# =======================================
# Instale com:
#   python -m pip install -r requirements.txt
#
# ATENÇÃO — Python 3.12.x obrigatório.

# --- Módulo de Escuta (STT + Wake Word offline) ---
SpeechRecognition==3.10.4
PyAudio==0.2.14
vosk==0.3.45
sounddevice==0.4.6

# --- Módulo de Fala (TTS) ---
edge-tts==6.1.12
pygame==2.6.1

# --- Módulo de Automação ---
requests==2.31.0

# --- WebSocket ---
websocket-client==1.7.0
websockets==12.0

# --- LLM: Claude API (opcional) ---
anthropic==0.40.0

# --- Obsidian Hot-reload ---
watchdog==4.0.0

# --- Visão Computacional ---
pillow==10.3.0

# --- Controle de Janelas ---
pygetwindow==0.0.9
pyautogui==0.9.54

# --- Notificações Windows ---
win10toast==0.9

# --- Google Calendar (opcional) ---
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.127.0

# --- Utilitários ---
python-dotenv==1.0.1
setuptools